+++ Stitch/Graph/Node/Port/ViewModel/NodeRowObserver/NodeRowObserverExtensions.swift
@@ -55,17 +55,17 @@ extension NodeRowObserver {
             self.hasLoopedValues = hasLoop
         }
                 
-        // Input AND output
-        // Update visual color data
-        self.allRowViewModels.forEach {
-            $0.updatePortColor()
-        }
-        
         switch Self.nodeIOType {
         case .input:
             self.inputPostProcessing(oldValues: oldValues, newValues: newValues)
         case .output:
-            self.updatePulsedOutputsForThisGraphStep()
+            self.outputPostProcessing()
+        }
+        
+        // Input AND output
+        // Update visual color data
+        self.allRowViewModels.forEach {
+            $0.updatePortColor()
         }
     }
